backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
openai>=1.0.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
gTTS>=2.5.0
reportlab>=4.0.0
Pillow>=10.0.0
# LangChain integration
langchain>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
